fix: resolve remaining compilation errors for release build
- DefaultAgentService.kt: Fixed combinedResultData type mismatch by ensuring consistent String return
- DeterministicUiExecutor.kt: Fixed recordRead to accept Long parameter
- EngineHolder.kt: Fixed LocalEngineLease constructor with non-null backendLabel
- ServerTokenStore.kt: Fixed verifyToken with null-safe providedBytes
- ChatScreen.kt: Fixed horizontalPad type mismatch and added missing braces for DropdownMenu
- ComposeChatActivity.kt: Added ExperimentalMaterial3WindowSizeClassApi annotation for calculateWindowSizeClass

+++ app/src/main/java/com/returngift/agent/agent/exec/DeterministicUiExecutor.kt
@@ -322,7 +322,7 @@ class DeterministicUiExecutor(
             return null
         }
         val tree = try { service.getScreenTree() } catch (_: Exception) { null } ?: return null
-        readGate.recordRead(tree.hashCode())
+        readGate.recordRead(tree.hashCode().toLong())
         return tree
     }
